Refactoring filenames and reset functions
