+++ components/anesthesia.py
@@ -4,18 +4,18 @@ import dash_html_components as html
 from dash.dependencies import Input, Output
 
 
-class Anaesthetic:
+class Anesthesia:
     STANDARD_VALUE = []
 
     @staticmethod
     def get_component():
         widget = dbc.FormGroup(
-            [dbc.Label("Anestesibedömning"), Anaesthetic._anaesthetic_checklist()]
+            [dbc.Label("Anestesibedömning"), Anesthesia._anesthesia_checklist()]
         )
         return widget
 
     @staticmethod
-    def _anaesthetic_checklist():
+    def _anesthesia_checklist():
         checklist = dbc.Checklist(
             options=[
                 {"label": "Ej klar", "value": "ek"},
@@ -28,12 +28,12 @@ class Anaesthetic:
         return checklist
 
     @staticmethod
-    def add_anaesthetic_callback(app):
+    def add_anesthesia_callback(app):
         @app.callback(
             Output(component_id="anaesthesia_checklist", component_property="value"),
             Input(component_id="reset_filter_button", component_property="n_clicks"),
         )
-        def reset_opTime(n_clicks):
-            return Anaesthetic.STANDARD_VALUE
+        def reset_component(n_clicks):
+            return Anesthesia.STANDARD_VALUE
 
         return app
