+++ components/asa.py
@@ -35,7 +35,7 @@ class Asa:
             Output(component_id="asa_checklist", component_property="value"),
             Input(component_id="reset_filter_button", component_property="n_clicks"),
         )
-        def reset_opTime(n_clicks):
+        def reset_component(n_clicks):
             return Asa.STANDARD_VALUE
 
         return app
