+++ components/op_code.py
@@ -58,7 +58,7 @@ class OpCode:
             Output(component_id="opCode_dropdown", component_property="value"),
             Input(component_id="reset_filter_button", component_property="n_clicks"),
         )
-        def reset_opTime(n_clicks):
+        def reset_component(n_clicks):
             return OpCode.STANDARD_VALUE_DD
 
         return app
