+++ components/statistics_code.py
@@ -47,7 +47,7 @@ class StatisticsCode:
             Output(component_id="statistics_dropdown", component_property="value"),
             Input(component_id="reset_filter_button", component_property="n_clicks"),
         )
-        def reset_opTime(n_clicks):
+        def reset_component(n_clicks):
             return StatisticsCode.STANDARD_VALUE
 
         return app
